requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
attrs
coloredlogs
discord.py
frozenlist
gphoto2
greenlet
humanfriendly
idna
mariadb
multidict
packaging
platformdirs
python-dateutil
pytz
six
SQLAlchemy
typing_extensions
yarl
